requirements.txt
fastapi
uvicorn
requests
python-dotenv
tweepy
langchain
langchain-community
chromadb
huggingface-hub